+++ backend/src/emails/service.py
