+++ backend/src/emails/templates/build_email/verify_email.html
